requirements.txt
Django
django-cassandra-engine
django-crispy-forms
